+++ src/contexts/Balances/Provider.tsx
@@ -4,11 +4,12 @@ import { useWallet } from "use-wallet";
 import { provider } from "web3-core";
 
 import { TGE1, ESCHUBQSLPAddress } from "constants/tokenAddresses";
-import { getBalance, getCurrentBlock } from "utils";
+import { getBalance, getCurrentBlock, getCoinBalance } from "utils";
 
 import Context from "./Context";
 
 const Provider: React.FC = ({ children }) => {
+  const [UBQBalance, setUBQBalance] = useState<BigNumber>();
   const [TGE1Balance, setTGE1Balance] = useState<BigNumber>();
   const [ESCHUBQLPBalance, setESCHUBQLPBalance] = useState<BigNumber>();
   const [CurrentBlock, setCurrentBlock] = useState("");
@@ -17,11 +18,17 @@ const Provider: React.FC = ({ children }) => {
 
   const fetchBalances = useCallback(
     async (userAddress: string, provider: provider) => {
-      const balances = await Promise.all([await getBalance(provider, TGE1, userAddress), await getBalance(provider, ESCHUBQSLPAddress, userAddress)]);
+      const balances = await Promise.all([
+        await getBalance(provider, TGE1, userAddress),
+        await getBalance(provider, ESCHUBQSLPAddress, userAddress),
+        await getCoinBalance(provider, userAddress),
+        // await getBalance(provider, "0x1FA6A37c64804C0D797bA6bC1955E50068FbF362", userAddress)
+      ]);
       setTGE1Balance(new BigNumber(balances[0]).dividedBy(new BigNumber(10).pow(18)));
       setESCHUBQLPBalance(new BigNumber(balances[1]).dividedBy(new BigNumber(10).pow(18)));
+      setUBQBalance(new BigNumber(balances[2]).dividedBy(new BigNumber(10).pow(18)));
     },
-    [setTGE1Balance, setESCHUBQLPBalance]
+    [setTGE1Balance, setESCHUBQLPBalance, setUBQBalance]
   );
 
   const fetchCurrentBlock = useCallback(
@@ -57,6 +64,7 @@ const Provider: React.FC = ({ children }) => {
       value={{
         TGE1Balance,
         ESCHUBQLPBalance,
+        UBQBalance,
         CurrentBlock,
       }}
     >