+++ src/contexts/Balances/types.ts
@@ -3,5 +3,6 @@ import BigNumber from "bignumber.js";
 export interface ContextValues {
   TGE1Balance?: BigNumber;
   ESCHUBQLPBalance?: BigNumber;
+  UBQBalance?: BigNumber;
   CurrentBlock?: string;
 }